- fixed delete on success
OODT-683

git-svn-id: https://svn.apache.org/repos/asf/oodt/trunk@1592885 13f79535-47bb-0310-9956-ffa450edef68

+++ pushpull/src/main/java/org/apache/oodt/cas/pushpull/retrievalsystem/RetrievalSetup.java
@@ -244,9 +244,9 @@ public class RetrievalSetup {
 
     private void movePropsFileToFinalDestination(PropFilesInfo pfi,
             File dirstructFile, String errorMsgs) throws IOException {
-        if (pfi.getDeleteOnSuccess()) {
-          dirstructFile.delete();
-          return;
+        if (pfi.getDeleteOnSuccess() && errorMsgs != null) {
+            dirstructFile.delete();
+            return;
         }
         File moveToDir = pfi.getFinalDestination(errorMsgs == null);
         moveToDir.mkdirs();
